Replace ServerConnectionDetails with RemoteServerConfig
It was more or less the same type as the stripped-down ServerConfig
so we can save some effort shuffling bytes between objects.

This also restructures the ConnectToServerDialog result handling
so that a single result is returned instead of having an accepted flag
and multiple result properties which enables simply annotating the
result as nullable instead of documenting that if accepted() returns
true then the result parameters won't be null, but if it's false then
they may be.